+++ changelog.md
@@ -1,5 +1,6 @@
 # 0.121
 
+- Update about page with new tutorial scenario
 - Bug fix: movement bug introduced by previous bridge fix
 - Promoted 000 (eh...  It's even enough for a tutorial, testing probably won't help)
 - Updated 004